requirements.txt
# Core AI/ML Libraries
langchain>=0.1.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0

# Vector Database and Embeddings
chromadb>=0.4.0
pinecone-client>=2.2.0
sentence-transformers>=2.2.0

# Web Framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# AI Models
anthropic>=0.18.0
langchain-anthropic>=0.1.0
pypdf2>=3.0.0
pyyaml>=6.0 